+++ ebong/src/app/component/LetterList.tsx
@@ -5,7 +5,7 @@ import { useQuery } from "@tanstack/react-query";
 import { letterClientApi } from "@/lib/client-api/letters";
 
 export const LetterList = () => {
-  const { data, error } = useQuery({
+  const { data } = useQuery({
     queryKey: ["letters"],
     queryFn: letterClientApi.getLetters,
   });
