+++ ebong/src/app/layout.tsx
@@ -1,7 +1,6 @@
 import { letterClientApi } from "@/lib/client-api/letters";
 import { voteProfileClientApi } from "@/lib/client-api/vote-profiles";
 import type { Metadata } from "next";
-import LettersProvider from "@/provider/letter-provider";
 import VoteProfilesProvider from "@/provider/vote-profile-provider";
 import "./globals.css";
 import localFont from "next/font/local";
@@ -23,20 +22,17 @@ export default async function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const letter_data = await letterClientApi.getLetters();
   const vote_profile_data = await voteProfileClientApi.getVoteProfiles();
 
   return (
     <html lang="en">
       <body className={omu.className}>
         <QueryProviders>
-          <LettersProvider initialLetters={letter_data.letters}>
-            <VoteProfilesProvider
-              initialVoteProfiles={vote_profile_data.vote_profiles}
-            >
-              {children}
-            </VoteProfilesProvider>
-          </LettersProvider>
+          <VoteProfilesProvider
+            initialVoteProfiles={vote_profile_data.vote_profiles}
+          >
+            {children}
+          </VoteProfilesProvider>
         </QueryProviders>
       </body>
     </html>
